+++ packages/redux/src/index.ts
@@ -1,30 +1,45 @@
 import { CuriConfig, Response } from '@curi/core';
-import { Store, Action } from 'redux';
+import { Store, Action as ReduxAction } from 'redux';
+import { Action as HickoryAction } from '@hickory/root';
 
 export const LOCATION_CHANGE = '@@curi/LOCATION_CHANGE';
 export const ADD_CURI = '@@curi/ADD_CURI';
 
-export interface ResponseAction extends Action {
+export interface ResponseAction extends ReduxAction {
   response: Response;
+  action: HickoryAction;
 }
 
-export interface CuriAction extends Action {
+export interface CuriAction extends ReduxAction {
   curi: CuriConfig;
 }
 
-export const responseReducer = (
-  state: Response,
-  action: Action
-): Response => {
-  return action.type === LOCATION_CHANGE
-    ? (<ResponseAction>action).response
-    : state === undefined ? null : state;
+export interface CuriState {
+  config: CuriConfig;
+  response: Response;
+  action: HickoryAction;
+}
+
+const INITIAL_STATE: CuriState = {
+  config: null,
+  response: null,
+  action: null
 };
 
-export const curiReducer = (state: CuriConfig, action: Action): CuriConfig => {
-  return action.type === ADD_CURI
-    ? (<CuriAction>action).curi
-    : state === undefined ? null : state;
+export const curiReducer = (state: CuriState = INITIAL_STATE, action: ReduxAction): CuriState => {
+  switch (action.type) {
+    case ADD_CURI:
+      return Object.assign({}, state, {
+        config: (<CuriAction>action).curi
+      });
+    case LOCATION_CHANGE:
+      return Object.assign({}, state, {
+        response: (<ResponseAction>action).response,
+        action: (<ResponseAction>action).action
+      });
+    default:
+      return state;
+  }
 };
 
 export const syncResponses = (store: Store<any>, curi: CuriConfig): void => {
@@ -33,10 +48,11 @@ export const syncResponses = (store: Store<any>, curi: CuriConfig): void => {
     curi
   });
 
-  curi.subscribe(response => {
+  curi.subscribe((response, action) => {
     store.dispatch({
       type: LOCATION_CHANGE,
-      response
+      response,
+      action
     });
   });
 };